Crate vizia_core

Vizia

Modules

animation

binding

Data binding provides a way to link views to model data so that view properties update when data changes.

context

Context types for retained state, used during view building, event handling, and drawing.

environment

A model for system specific state which can be accessed by any model or view.

events

Events for communicating state change to views and models.

icons

A collection of built-in SVG icons.

input

Types used for handling input events such as mouse and keyboard.

layout

Vizia uses morphorm for layout.

localization

Provides types for adapting an application to a particular language or regional peculiarities.

model

Models are used to store application data and can be bound to by views to visually display the data.

modifiers

Methods on views for changing their properties or for adding actions.

resource

Resource management for fonts, themes, images, and translations.

style

Styling determines the appearance of a view.

util

Helper utilities

vg

Contains types and functions used for custom drawing within views. This is a re-export of skia-safe.

view

Views are used to visually present model data and to act as controls which, when interacted with, send events to mutate model data.

views

Built-in views provided by vizia.

window

Window management and system events.

Macros

impl_res_clone

A macro for implementing the Res trait for Clone types.

impl_res_simple

A macro for implementing the Res trait for simple Copy types.

include_style

A macro which parses CSS from a file at runtime in debug mode, and includes the file in the binary in release mode.