Trait vizia_core::binding::LensExt

pub trait LensExt: Lens {
    // Provided methods
    fn get<C: DataContext>(&self, cx: &C) -> Self::Target
       where Self::Target: Clone { ... }
    fn get_fallible<C: DataContext>(&self, cx: &C) -> Option<Self::Target>
       where Self::Target: Clone { ... }
    fn or<Other>(self, other: Other) -> OrLens<Self, Other>
       where Other: Lens<Target = bool>,
             Self: Lens<Target = bool> { ... }
    fn and<Other>(self, other: Other) -> AndLens<Self, Other>
       where Other: Lens<Target = bool>,
             Self: Lens<Target = bool> { ... }
    fn then<Other>(self, other: Other) -> Then<Self, Other>
       where Other: Lens<Source = Self::Target> { ... }
    fn index<T>(self, index: usize) -> Index<Self, T>
       where T: 'static,
             Self::Target: Deref<Target = [T]> { ... }
    fn map<O: 'static, F: 'static + Fn(&Self::Target) -> O>(
        self,
        map: F
    ) -> Map<Self, O> { ... }
    fn map_ref<O: 'static, F: 'static + Fn(&Self::Target) -> &O>(
        self,
        map: F
    ) -> MapRef<Self, O> { ... }
    fn unwrap<T: 'static>(self) -> Then<Self, UnwrapLens<T>>
       where Self: Lens<Target = Option<T>> { ... }
    fn into_lens<T: 'static>(self) -> Then<Self, IntoLens<Self::Target, T>>
       where Self::Target: Clone + Into<T> { ... }
}

Helpers for constructing more complex Lenses.

Provided Methods

fn get<C: DataContext>(&self, cx: &C) -> Self::Targetwhere Self::Target: Clone,

Retrieve a copy of the lensed data from context. This will clone the data.

Example

let value = lens.get(cx);

fn get_fallible<C: DataContext>(&self, cx: &C) -> Option<Self::Target>where Self::Target: Clone,

fn or<Other>(self, other: Other) -> OrLens<Self, Other>where Other: Lens<Target = bool>, Self: Lens<Target = bool>,

fn and<Other>(self, other: Other) -> AndLens<Self, Other>where Other: Lens<Target = bool>, Self: Lens<Target = bool>,

fn then<Other>(self, other: Other) -> Then<Self, Other>where Other: Lens<Source = Self::Target>,

Used to construct a lens to some data contained within some other lensed data.

Example

Binds a label to other_data, which is a field of a struct SomeData, which is a field of the root AppData model:

Binding::new(cx, AppData::some_data.then(SomeData::other_data), |cx, data|{

});

fn index<T>(self, index: usize) -> Index<Self, T>where T: 'static, Self::Target: Deref<Target = [T]>,

fn map<O: 'static, F: 'static + Fn(&Self::Target) -> O>( self, map: F ) -> Map<Self, O>

fn map_ref<O: 'static, F: 'static + Fn(&Self::Target) -> &O>( self, map: F ) -> MapRef<Self, O>

fn unwrap<T: 'static>(self) -> Then<Self, UnwrapLens<T>>where Self: Lens<Target = Option<T>>,

fn into_lens<T: 'static>(self) -> Then<Self, IntoLens<Self::Target, T>>where Self::Target: Clone + Into<T>,