vizia_core::context

Trait TreeProps

pub trait TreeProps {
    // Required methods
    fn parent(&self) -> Entity;
    fn first_child(&self) -> Entity;
    fn parent_window(&self) -> Entity;
}

Trait for querying properties of the tree from a context.

Required Methods

fn parent(&self) -> Entity

Returns the entity id of the parent of the current view.

fn first_child(&self) -> Entity

Returns the entity id of the first_child of the current view.

fn parent_window(&self) -> Entity

Returns the entity id of the parent window of the current view.