Trait vizia_core::events::ViewHandler

pub trait ViewHandler: Any {
    fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event);
    fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas);
    fn as_any_ref(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;

    fn element(&self) -> Option<&'static str> { ... }
}

Required Methods

fn event(&mut self, cx: &mut EventContext<'_>, event: &mut Event)

fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas)

fn as_any_ref(&self) -> &dyn Any

fn as_any_mut(&mut self) -> &mut dyn Any

Provided Methods

fn element(&self) -> Option<&'static str>

Implementations

impl dyn ViewHandler

pub fn downcast_ref<T: Any>(&self) -> Option<&T>

Attempt to cast a view handler to an immutable reference to the specified type.

pub fn downcast_mut<T: Any>(&mut self) -> Option<&mut T>

Attempt to cast a view handler to a mutable reference to the specified type.