Module vizia_core::input

Types used for handling input events such as mouse and keyboard.

Structs

Keymap

A keymap that associates key chords with actions.

KeymapEntry

An entry inside of a Keymap.

Modifiers

The state of the modifier keys.

MouseButtonData

Data which describes the current state of a mouse button.

MouseState

The current state of the mouse cursor and buttons.

Enums

Code

Code is the physical position of a key.

Key

Key represents the meaning of a keypress.

KeymapEvent

An event used to interact with a Keymap at runtime.

MouseButton

A mouse button.