vizia_core::modifiers
Trait AccessibilityModifiers

pub trait AccessibilityModifiers: Modifiable {
    // Provided methods
    fn role(self, role: Role) -> Self { ... }
    fn name<U: ToStringLocalized>(self, name: impl Res<U>) -> Self { ... }
    fn live(self, live: Live) -> Self { ... }
    fn hidden<U: Into<bool>>(self, hidden: impl Res<U>) -> Self { ... }
    fn numeric_value<U: Into<f64>>(self, value: impl Res<U>) -> Self { ... }
    fn text_value<U: ToStringLocalized>(self, value: impl Res<U>) -> Self { ... }
}

Modifiers for changing the accessibility properties of a view.

Provided Methods

fn role(self, role: Role) -> Self

Sets the accessibility role of the view.

fn name<U: ToStringLocalized>(self, name: impl Res<U>) -> Self

Sets the accessibility name of the view.

fn live(self, live: Live) -> Self

Sets whether the view should act as an accessibility live region.

fn hidden<U: Into<bool>>(self, hidden: impl Res<U>) -> Self

Sets whether the view should be hidden from accessibility.

fn numeric_value<U: Into<f64>>(self, value: impl Res<U>) -> Self

Sets the accessibility numeric value for the view.

fn text_value<U: ToStringLocalized>(self, value: impl Res<U>) -> Self

Sets the accessibility text value for the view.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.