vizia_core::modifiers
Trait TextModifiers

pub trait TextModifiers: Modifiable {

    // Provided methods
    fn text<T: ToStringLocalized>(self, value: impl Res<T>) -> Self { ... }
    fn font_family<U: Into<Vec<FamilyOwned>>>(self, value: impl Res<U>) -> Self { ... }
    fn font_weight<U: Into<FontWeight>>(self, value: impl Res<U>) -> Self { ... }
    fn font_slant<U: Into<FontSlant>>(self, value: impl Res<U>) -> Self { ... }
    fn font_width<U: Into<FontWidth>>(self, value: impl Res<U>) -> Self { ... }
    fn font_variation_settings<U: Into<Vec<FontVariation>>>(
        self,
        value: impl Res<U>,
    ) -> Self { ... }
    fn color<U: Clone + Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn font_size<U: Into<FontSize>>(self, value: impl Res<U>) -> Self { ... }
    fn caret_color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn selection_color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn text_wrap<U: Into<bool>>(self, value: impl Res<U>) -> Self { ... }
    fn text_align<U: Into<TextAlign>>(self, value: impl Res<U>) -> Self { ... }
    fn text_overflow<U: Into<TextOverflow>>(self, value: impl Res<U>) -> Self { ... }
    fn line_clamp<U: Into<LineClamp>>(self, value: impl Res<U>) -> Self { ... }
    fn text_decoration_line<U: Into<TextDecorationLine>>(
        self,
        value: impl Res<U>,
    ) -> Self { ... }
    fn text_stroke_width<U: Into<Length>>(self, value: impl Res<U>) -> Self { ... }
    fn text_stroke_style<U: Into<TextStrokeStyle>>(
        self,
        value: impl Res<U>,
    ) -> Self { ... }

}

Modifiers for changing the text properties of a view.

Provided Methods

fn text<T: ToStringLocalized>(self, value: impl Res<T>) -> Self

Sets the text content of the view.

fn font_family<U: Into<Vec<FamilyOwned>>>(self, value: impl Res<U>) -> Self

Sets the font that should be used by the view.

The font name refers to the name assigned when the font is added to context.

fn font_weight<U: Into<FontWeight>>(self, value: impl Res<U>) -> Self

Sets the font weight that should be used by the view.

fn font_slant<U: Into<FontSlant>>(self, value: impl Res<U>) -> Self

Sets the font style that should be used by the view.

fn font_width<U: Into<FontWidth>>(self, value: impl Res<U>) -> Self

Sets the font stretch that should be used by the view if the font supports it.

fn font_variation_settings<U: Into<Vec<FontVariation>>>( self, value: impl Res<U>, ) -> Self

Sets the font variation settings that should be used by the view.

fn color<U: Clone + Into<Color>>(self, value: impl Res<U>) -> Self

Sets the text color of the view.

fn font_size<U: Into<FontSize>>(self, value: impl Res<U>) -> Self

Sets the font size of the view.

fn caret_color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the ext caret color of the view.

fn selection_color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the color used to highlight selected text within the view.

fn text_wrap<U: Into<bool>>(self, value: impl Res<U>) -> Self

Sets whether the text of the view should be allowed to wrap.

fn text_align<U: Into<TextAlign>>(self, value: impl Res<U>) -> Self

Sets the horizontal alignment of text within the view.

fn text_overflow<U: Into<TextOverflow>>(self, value: impl Res<U>) -> Self

Sets the text overflow.

fn line_clamp<U: Into<LineClamp>>(self, value: impl Res<U>) -> Self

Sets the max number of .

fn text_decoration_line<U: Into<TextDecorationLine>>( self, value: impl Res<U>, ) -> Self

Sets the max number of .

fn text_stroke_width<U: Into<Length>>(self, value: impl Res<U>) -> Self

Sets the width of the text stroke. This sets Skia’s skia_safe::textlayout::TextStyle’s foreground skia_safe::Paint to draw a stroke on the text.

See also Self::text_stroke_style.

fn text_stroke_style<U: Into<TextStrokeStyle>>(self, value: impl Res<U>) -> Self

Sets the paint style of the text stroke. You can either draw text with a stroke, or just the stroke outline.

If you desire to paint the text stroke in a separate colour to the fill property, Skia does not seem to support this (see the discussion on the PR to add text stroke).

A workaround would involve a ZStack to overlay two copies of the same text, one with a stroke and one without. These can have two different colours. See also the Flutter documentation on achieving this.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.