Struct vizia_core::prelude::EventContext

pub struct EventContext<'a> {
    pub style: &'a mut Style,
    pub cache: &'a CachedData,
    pub draw_cache: &'a DrawCache,
    pub tree: &'a Tree<Entity>,
    pub resource_manager: &'a ResourceManager,
    pub text_context: &'a mut TextContext,
    pub modifiers: &'a Modifiers,
    pub mouse: &'a MouseState<Entity>,
    /* private fields */
}

Fields
style: &'a mut Stylecache: &'a CachedDatadraw_cache: &'a DrawCachetree: &'a Tree<Entity>resource_manager: &'a ResourceManagertext_context: &'a mut TextContextmodifiers: &'a Modifiersmouse: &'a MouseState<Entity>
Implementations

impl<'a> EventContext<'a>

pub fn new(cx: &'a mut Context) -> Self

pub fn resolve_entity_identifier(&self, identity: &str) -> Option<Entity>

Finds the entity that identifier identifies

pub fn current(&self) -> Entity

pub fn add_listener<F, W>(&mut self, listener: F)where
W: View,
F: 'static + Fn(&mut W, &mut EventContext<'_>, &mut Event),

Add a listener to an entity.

A listener can be used to handle events which would not normally propagate to the entity. For example, mouse events when a different entity has captured them. Useful for things like closing a popup when clicking outside of its bounding box.

pub fn set_active(&mut self, active: bool)

Set the active state for the current entity.

pub fn capture(&mut self)

Capture mouse input for the current entity.

pub fn release(&mut self)

Release mouse input capture for current entity.

pub fn focus_with_visibility(&mut self, focus_visible: bool)

Sets application focus to the current entity with the specified focus visiblity

pub fn focus(&mut self)

Sets application focus to the current entity using the previous focus visibility.

pub fn hovered(&self) -> Entity

Return the currently hovered entity.

pub fn focused(&self) -> Entity

Return the currently focused entity.

pub fn is_disabled(&self) -> bool

Returns true if the current entity is disabled.

pub fn is_over(&self) -> bool

Returns true if the mouse cursor is over the current entity.

pub fn lock_cursor_icon(&mut self)

Prevents the cursor icon from changing until the lock is released.

pub fn unlock_cursor_icon(&mut self)

Releases any cursor icon lock, allowing the cursor icon to be changed.

pub fn is_cursor_icon_locked(&self) -> bool

Returns true if the cursor icon is locked.

pub fn set_hover(&mut self, flag: bool)

Sets the hover flag of the current entity.

pub fn set_checked(&mut self, flag: bool)

Sets the checked flag of the current entity.

pub fn set_selected(&mut self, flag: bool)

Sets the checked flag of the current entity.

pub fn get_clipboard(
&mut self
) -> Result<String, Box<dyn Error + Send + Sync + 'static>>

Get the contents of the system clipboard. This may fail for a variety of backend-specific reasons.

pub fn set_clipboard(
&mut self,
text: String
) -> Result<(), Box<dyn Error + Send + Sync + 'static>>

Set the contents of the system clipboard. This may fail for a variety of backend-specific reasons.

pub fn toggle_class(&mut self, class_name: &str, applied: bool)

pub fn play_animation(&mut self, animation: Animation)

pub fn environment(&self) -> &Environment

pub fn needs_redraw(&mut self)

pub fn needs_relayout(&mut self)

pub fn reload_styles(&mut self) -> Result<(), Error>

pub fn spawn<F>(&self, target: F)where
F: 'static + Send + FnOnce(&mut ContextProxy),

pub fn dpi_factor(&self) -> f32

The window’s DPI factor. This includes both HiDPI scaling and the user scale factor.

pub fn window_size(&self) -> WindowSize

The window’s size in logical pixels, before user_scale_factor() gets applied to it. If this value changed during a frame then the window will be resized and a WindowEvent::GeometryChanged will be emitted.

pub fn set_window_size(&mut self, new_size: WindowSize)

Change the window size. A WindowEvent::GeometryChanged will be emitted when the window has actually changed in size.

pub fn user_scale_factor(&self) -> f64

A scale factor used for uniformly scaling the window independently of any HiDPI scaling. window_size gets multplied with this factor to get the actual logical window size. If this changes during a frame, then the window will be resized at the end of the frame and a WindowEvent::GeometryChanged will be emitted. This can be initialized using [WindowDescription::user_scale_factor][crate::WindowDescription::user_scale_factor].

pub fn set_user_scale_factor(&mut self, new_factor: f64)

Change the user scale factor size. A WindowEvent::GeometryChanged will be emitted when the window has actually changed in size.