Trait vizia_core::prelude::Lens

pub trait Lens: 'static + Clone {
    type Source;
    type Target;

    fn view<O, F: FnOnce(Option<&Self::Target>) -> O>(
        &self, 
        source: &Self::Source, 
        map: F
    ) -> O;
}

A Lens allows the construction of a reference to a piece of some data, e.g. a field of a struct.

When deriving the Lens trait on a struct, the derive macro constructs a static type which implements the Lens trait for each field. The view() method takes a reference to the struct type as input and outputs a reference to the field. This provides a way to specify a binding to a specific field of some application data.

This trait is part of the prelude.

Required Associated Types

type Source

type Target

Required Methods

fn view<O, F: FnOnce(Option<&Self::Target>) -> O>(
&self,
source: &Self::Source,
map: F
) -> O