Module vizia_core::state

Data Binding

Binding provides reactivity to a vizia application. Rather than sending events back and forth between widgets to update local widget data, widgets can instead bind to application data.

Example

Fist we declare the data for our application. The Lens trait has been derived for the data, which allows us to bind to fields of the struct:

#[derive(Lens)]
struct AppData {
    count: i32,
}

Next we’ll declare some events which will be sent by widgets to modify the app data. Data binding in vizia is one-way, events are sent up the tree to the app data to mutate it and updated values are sent to observers, such as a Binding view.

enum AppEvent {
    Increment,
    Decrement,
}

Next we implement the Model trait on our app data, which allows us to modify the data in response to an Event:

impl Model for AppData {
    fn event(&mut self, cx: &mut Context, event: &mut Event) {
        event.map(|app_event, _| match app_event {
            AppEvent::Increment => {
                self.count += 1;
            }

            AppEvent::Decrement => {
                self.count -= 1;
            }
        });
    }
}

This trait also allows data to be built into the application Tree:

fn main() {
    Application::new(|cx|{
        AppData {
            count: 0,
        }.build(cx);
    }).run();  
}

A Binding view is one way in which data can be used by widgets. A Lens is used to determine what data the binding should react to:

fn main() {
    Application::new(|cx|{
        AppData {
            count: 0,
        }.build(cx);

        Binding::new(cx, AppData::count, |cx, count|{
            Label::new(cx, &count.get(cx).to_string());
        });
    }).run();
}

The second parameter to the Binding view is a Lens, allowing us to bind to some field of the application data. The third parameter is a closure which provides the context and the lens, which can be used to retrieve the bound data using the .get() method, which takes the Context as an argument.

Now when the data is modified by another widget, the label will update, for example:

fn main() {
    Application::new(|cx|{
        AppData {
            count: 0,
        }.build(cx);

        Binding::new(cx, AppData::count, |cx, count|{
            Label::new(cx, &count.get(cx).to_string());
        });

        Button::new(cx, |cx| cx.emit(AppEvent::Increment), |cx|{
            Label::new(cx, "Increment")
        });

        Button::new(cx, |cx| cx.emit(AppEvent::Increment), |cx|{
            Label::new(cx, "Decrement")
        });
    }).run();
}

Note, the checkbox does not need to be bound to the data to send an event to it. By default events will propagate up the tree.

Structs

BasicStore

Binding

A binding view which rebuilds its contents when its observed data changes.

Index

IntoLens

Map

MapShallow

RatioLens

StaticLens

StoreId

Then

Lens composed of two lenses joined together

UnwrapLens

Enums

ModelOrView

Traits

Bindable

BindableExt

Data

A trait for any type which can be bound to, i.e. can be cached and compared against previous versions.

Lens

A Lens allows the construction of a reference to a piece of some data, e.g. a field of a struct.

LensExt

Helpers for constructing more complex Lenses.

Model

A trait implemented by application data in order to mutate in response to events.

ModelData

Res

A trait for types that can automatically resolve into other types, with or without consulting the Context.

Store

StoreHandler