Trait vizia_core::state::LensExt

pub trait LensExt: Lens {
    fn get<C: DataContext>(&self, cx: &C) -> Self::Target
    where
        Self::Target: Clone,
    { ... }

    fn get_fallible<C: DataContext>(&self, cx: &C) -> Option<Self::Target>
    where
        Self::Target: Clone,
    { ... }

    fn then<Other>(self, other: Other) -> Then<Self, Other>
    where
        Other: Lens<Source = Self::Target>,
        Self: Sized,
    { ... }

    fn index<T>(self, index: usize) -> Then<Self, Index<Self::Target, T>>
    where
        T: 'static,
        Self::Target: Deref<Target = [T]>,
    { ... }

    fn map<G, B: 'static>(self, get: G) -> Then<Self, Map<Self::Target, B>>
    where
        G: 'static + Fn(&Self::Target) -> B,
    { ... }

    fn unwrap<T: 'static>(self) -> Then<Self, UnwrapLens<T>>
    where
        Self: Lens<Target = Option<T>>,
    { ... }

    fn into_lens<T: 'static>(self) -> Then<Self, IntoLens<Self::Target, T>>
    where
        Self::Target: Clone + Into<T>,
    { ... }
}

Helpers for constructing more complex Lenses.

This trait is par tof the prelude.

Provided methods

fn get<C: DataContext>(&self, cx: &C) -> Self::Target where
Self::Target: Clone,

Retrieve a copy of the lensed data from context.

Example

let value = lens.get(cx);

fn get_fallible<C: DataContext>(&self, cx: &C) -> Option<Self::Target> where
Self::Target: Clone,

fn then<Other>(self, other: Other) -> Then<Self, Other> where
Other: Lens<Source = Self::Target>,
Self: Sized,

Used to construct a lens to some data contained within some other lensed data.

Example

Binds a label to other_data, which is a field of a struct SomeData, which is a field of the root AppData model:

Binding::new(cx, AppData::some_data.then(SomeData::other_data), |cx, data|{

});

fn index<T>(self, index: usize) -> Then<Self, Index<Self::Target, T>> where
T: 'static,
Self::Target: Deref<Target = [T]>,

fn map<G, B: 'static>(self, get: G) -> Then<Self, Map<Self::Target, B>> where
G: 'static + Fn(&Self::Target) -> B,

fn unwrap<T: 'static>(self) -> Then<Self, UnwrapLens<T>> where
Self: Lens<Target = Option<T>>,

fn into_lens<T: 'static>(self) -> Then<Self, IntoLens<Self::Target, T>> where
Self::Target: Clone + Into<T>,