Enum vizia_core::style::TextAlign

pub enum TextAlign {
    Start,
    End,
    Left,
    Right,
    Center,
    Justify,
}

Determines how overflowed content that is not displayed should be signaled to the user.

Variants

Start

The same as left if direction is left-to-right and right if direction is right-to-left.

End

The same as right if direction is left-to-right and left if direction is right-to-left.

Left

The inline contents are aligned to the left edge of the line box.

Right

The inline contents are aligned to the right edge of the line box.

Center

The inline contents are centered within the line box.

Justify

The inline contents are justified. Text should be spaced to line up its left and right edges to the left and right edges of the line box, except for the last line.