Module vizia_core::tree

Functions

focus_backward

focus_forward

is_focusable

Is the entity focusable - some focusable entities are not in the tab order.

is_navigatable

Should the user be able to navigate to the entity with tab?