vizia_core::vg
Module gpu

Modules

backend_formats

backend_render_targets

backend_textures

direct_contexts

ganesh

gl

interfaces

surfaces

Structs

Budgeted

ContextOptions

DirectContextId

DriverBugWorkarounds

FlushInfo

SubmitInfo

Enums

BackendAPI

BackendApi

Mipmapped

Origin

Protected

PurgeResourceOptions

Renderable

SemaphoresSubmitted

SurfaceOrigin

SyncCpu

Type Aliases

BackendFormat

BackendRenderTarget

BackendTexture

DirectContext

MutableTextureState

RecordingContext

YUVABackendTextureInfo

A description of a set BackendTextures that hold the planar data described by a YUVAInfo.

YUVABackendTextures

A set of BackendTextures that hold the planar data for an image described a YUVAInfo.