vizia_core::vg::gpu::ganesh::images
Function get_backend_texture_from_image

pub fn get_backend_texture_from_image(
    image: &RCHandle<SkImage>,
    flush_pending_gr_context_io: bool,
) -> Option<(RefHandle<GrBackendTexture>, GrSurfaceOrigin)>

Retrieves the existing backend texture. If Image is not a Ganesh-backend texture image or otherwise does not have such a texture, false is returned. Otherwise, returned will be set to the image’s texture.

If flush_pending_gr_context_io is true, completes deferred I/O operations. If origin in not None, copies location of content drawn into Image.

out_texture - Will be set to the underlying texture of the image if non-null.

flush_pending_gr_context_io - flag to flush outstanding requests

origin - Will be set to the origin orientation of the image if non-null.

Returns: None if a Ganesh backend texture cannot be retrieved.