vizia_core::vg::image_filters
Function blend

pub fn blend(
    mode: impl Into<RCHandle<SkBlender>>,
    background: impl Into<Option<RCHandle<SkImageFilter>>>,
    foreground: impl Into<Option<RCHandle<SkImageFilter>>>,
    crop_rect: impl Into<CropRect>,
) -> Option<RCHandle<SkImageFilter>>

This filter takes an crate::BlendMode and uses it to composite the two filters together.

blender - The blender that defines the compositing operation

background - The Dst pixels used in blending, if null the source bitmap is used.

foreground - The Src pixels used in blending, if null the source bitmap is used.

`crop_rect`` Optional rectangle to crop input and output.