vizia_core::vg
Module path

Structs

Iter

Iterates through verb array, and associated Point array and conic weight. Provides options to treat open contours as closed, and to ignore degenerate data.

RawIterDeprecated

SegmentMask

Enums

AddPathMode

ArcSize

Direction

FillType

Verb

Type Aliases

Path