Struct vizia_core::vg::rgb::alt::BGRA

#[repr(C)]
pub struct BGRA<ComponentType, AlphaComponentType = ComponentType> {
    pub b: ComponentType,
    pub g: ComponentType,
    pub r: ComponentType,
    pub a: AlphaComponentType,
}

BGR+A

Fields
b: ComponentType

Blue first

g: ComponentType

Green

r: ComponentType

Red

a: AlphaComponentType

Alpha last

Implementations

impl<T> BGRA<T, T>

pub const fn new(r: T, g: T, b: T, a: T) -> BGRA<T, T>

👎 Deprecated:
This function has a misleading order of arguments. Use BGRA{} literal instead

Convenience function for creating a new pixel Warning: The order of arguments is R,G,B,A

impl<T, A> BGRA<T, A>

pub const fn new_alpha(r: T, g: T, b: T, a: A) -> BGRA<T, A>

👎 Deprecated:
This function has a misleading order of arguments. Use BGRA{} literal instead

Convenience function for creating a new pixel Warning: The order of arguments is R,G,B,A

impl<T, A> BGRA<T, A>

pub fn rgb_mut(&mut self) -> &mut BGR<T>

👎 Deprecated:
This function will change. Use bgr_mut()

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

pub fn bgr_mut(&mut self) -> &mut BGR<T>

Provide a mutable view of only RGB components (leaving out alpha). Useful to change color without changing opacity.

impl<T, A> BGRA<T, A> where
T: Clone,

pub fn rgb(&self) -> BGR<T>

👎 Deprecated:
This function will change. Use bgr()

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T> BGRA<T, T> where
T: Clone,

pub fn iter(&self) -> Cloned<Iter<'_, T>>

Iterate over all components (length=4)

impl<T, A> BGRA<T, A> where
T: Clone,

pub fn bgr(&self) -> BGR<T>

Copy RGB components out of the RGBA struct

Note: you can use .into() to convert between other types

impl<T, A> BGRA<T, A> where
T: Copy,
A: Clone,

pub fn map_rgb<F, U, B>(&self, f: F) -> BGRA<U, B> where
F: FnMut(T) -> U,
U: Clone,
B: From<A> + Clone,

Create new RGBA with the same alpha value, but different RGB values

pub fn alpha(&self, a: A) -> BGRA<T, A>

Create a new RGBA with the new alpha value, but same RGB values

pub fn map_alpha<F, B>(&self, f: F) -> BGRA<T, B> where
F: FnOnce(A) -> B,

Create a new RGBA with a new alpha value created by the callback. Allows changing of the type used for the alpha channel.