Trait vizia_core::vg::rgb::Zeroable

pub unsafe trait Zeroable: Sized {
    // Provided method
    fn zeroed() -> Self { ... }
}

Trait for types that can be safely created with zeroed.

An all-zeroes value may or may not be the same value as the Default value of the type.

Safety

Your type must be inhabited (eg: no Infallible).

Your type must be allowed to be an “all zeroes” bit pattern (eg: no NonNull<T>).

Features

Some impls are feature gated due to the MSRV policy:

MaybeUninit<T> was not available in 1.34.0, but is available under the zeroable_maybe_uninit feature flag.

Atomic* types require Rust 1.60.0 or later to work on certain platforms, but is available under the zeroable_atomics feature flag.

[T; N] for arbitrary N requires the min_const_generics feature flag.

Provided Methods

fn zeroed() -> Self

Calls zeroed.

This is a trait method so that you can write MyType::zeroed() in your code. It is a contract of this trait that if you implement it on your type you must not override this method.

Object Safety

This trait is not object safe.