vizia_core::vg

Module shaders

Functions

blend

color

color_in_space

coord_clamp

empty

fractal_noise

image

Create an Shader that will sample the ‘image’. This is equivalent to Image::to_shader.

linear_gradient

Returns a shader that generates a linear gradient between the two specified points.

radial_gradient

Returns a shader that generates a radial gradient given the center and radius.

raw_image

Create an Shader that will sample ‘image’ with minimal processing. This is equivalent to Image::to_raw_shader.

sweep_gradient

Returns a shader that generates a sweep gradient given a center.

turbulence

two_point_conical_gradient

Returns a shader that generates a conical gradient given two circles.