vizia_core::view

Struct Handle

pub struct Handle<'a, V> { /* private fields */ }

A handle to a view which has been built into the tree.

Implementations

impl<V> Handle<'_, V>

pub fn entity(&self) -> Entity

Returns the [Entity] id of the view.

pub fn context(&mut self) -> &mut Context

Returns a mutable reference to the context.

pub fn parent(&self) -> Entity

Returns the entity id of the parent view.

pub fn lock_focus_to_within(self) -> Self

Stop the user from tabbing out of a subtree, which is useful for modal dialogs.

pub fn modify<F>(self, f: F) -> Self
where F: FnOnce(&mut V), V: 'static,

Mody the internal data of the view.

pub fn on_build<F>(self, callback: F) -> Self
where F: Fn(&mut EventContext<'_>),

Callback which is run when the view is built/rebuilt.

pub fn bind<R, T, F>(self, signal: R, closure: F) -> Self
where R: SignalGet<T> + 'static, T: Clone + 'static, F: 'static + Fn(Handle<'_, V>),

Creates a binding to the given resource and provides a closure which can be used to mutate the view through a handle.

pub fn needs_relayout(&mut self)

Marks the view as needing a relayout.

pub fn needs_restyle(&mut self)

Marks the view as needing a restyle.

pub fn needs_redraw(&mut self)

Marks the view as needing a redraw.

pub fn bounds(&self) -> BoundingBox

Returns the bounding box of the view.

pub fn scale_factor(&self) -> f32

Returns the scale factor of the device.

impl Handle<'_, Accordion>

pub fn open(self, indices: impl Res<Vec<usize>> + 'static) -> Self

Sets which sections are open by index.

pub fn on_toggle<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, usize, bool),

Set a callback that fires when a section is toggled open or closed.

The callback receives the section index and the desired next open state. Use this to update the external signal passed to open().

impl Handle<'_, AvatarGroup>

pub fn max_visible(self, max_visible: impl Res<usize> + 'static) -> Self

Limits the number of visible avatars in a group and adds a final overflow avatar with a +N label when more avatars are present.

impl Handle<'_, Badge>

pub fn placement<U: Into<BadgePlacement> + Clone + 'static>( self, placement: impl Res<U> + 'static, ) -> Self

Sets the placement of a badge relative to its parent. Accepts a value or signal of type BadgePlacement.

impl Handle<'_, ButtonGroup>

pub fn vertical(self, is_vertical: impl Res<bool>) -> Self

Sets whether the button group is in vertical orientation.

impl Handle<'_, Calendar>

pub fn on_select<F: 'static + Fn(&mut EventContext<'_>, NaiveDate)>( self, callback: F, ) -> Self

Set the callback triggered when a date is selected from the Calendar view.

impl Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, value)
    .on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, Chip>

pub fn on_close( self, callback: impl 'static + Fn(&mut EventContext<'_>) + Send + Sync, ) -> Self

Set the callback triggered when the close button of the chip is pressed. The chip close button is not displayed by default. Setting this callback causes the close button to be displayed.

pub fn variant<U: Into<ChipVariant> + Clone + 'static>( self, variant: impl Res<U> + 'static, ) -> Self

Selects the style variant to be used by the chip. Accepts a value or signal of type ChipVariant.

Example

Chip::new(cx, "Chip")
    .variant(ChipVariant::Filled);

impl Handle<'_, Collapsible>

pub fn open(self, open: impl Res<bool> + 'static) -> Self

Set the open state of the collapsible view.

pub fn on_toggle( self, callback: impl Fn(&mut EventContext<'_>, bool) + 'static, ) -> Self

Set the callback which is triggered when the open state changes.

impl<L, S, T> Handle<'_, ComboBox<L, S, T>>
where L: SignalGet<Vec<T>> + 'static, T: 'static + Clone + ToStringLocalized, S: SignalGet<usize> + 'static,

pub fn on_select<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when an item is selected from the ComboBox list.

impl Handle<'_, Dropdown>

pub fn placement(self, placement: impl Res<Placement> + 'static) -> Self

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn show_arrow(self, show_arrow: impl Res<bool> + 'static) -> Self

Sets whether the popup should include an arrow. Defaults to true.

pub fn arrow_size<U: Into<Length> + Clone + 'static>( self, size: impl Res<U> + 'static, ) -> Self

Sets the size of the popup arrow, or gap if the arrow is hidden.

pub fn should_reposition(self, flag: impl Res<bool> + 'static) -> Self

Set to whether the popup should reposition to always be visible.

impl<T: Res<f32> + 'static> Handle<'_, Knob<T>>

pub fn on_change<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the knob value is changed.

impl Handle<'_, ArcTrack>

pub fn value<R: Res<f32>>(self, value: R) -> Self

impl Handle<'_, TickKnob>

pub fn value<R: Res<f32>>(self, value: R) -> Self

impl Handle<'_, Label>

pub fn describing(self, entity_identifier: impl Into<String>) -> Self

Which form element does this label describe.

Examples

Checkbox::new(cx, value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue)).id("checkbox_identifier");
Label::new(cx, "hello").describing("checkbox_identifier");

impl Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, value)
    .on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, Rating>

pub fn on_change<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, u32),

Set the callback which is triggered when the rating changes.

impl Handle<'_, Resizable>

pub fn on_reset<F>(self, on_reset: F) -> Self
where F: Fn(&mut EventContext<'_>) + 'static,

Sets a callback to be called when the stack is reset, i.e. when the resize handle is double-clicked.

impl Handle<'_, Scrollbar>

pub fn scroll_to_cursor( self, scroll_to_cursor: impl Res<bool> + 'static, ) -> Self

Sets whether the scrollbar should move to the cursor when pressed.

impl Handle<'_, ScrollView>

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + 'static + Send + Sync, ) -> Self

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_to_cursor( self, scroll_to_cursor: impl Res<bool> + 'static, ) -> Self

Sets whether the scrollbar should move to the cursor when pressed.

pub fn scroll_x(self, scrollx: impl Res<f32> + 'static) -> Self

Set the horizontal scroll position of the ScrollView. Accepts a value or signal of type f32 between 0 and 1.

pub fn scroll_y(self, scrolly: impl Res<f32> + 'static) -> Self

Set the vertical scroll position of the ScrollView. Accepts a value or signal of type f32 between 0 and 1.

pub fn show_horizontal_scrollbar(self, flag: impl Res<bool> + 'static) -> Self

Sets whether the horizontal scrollbar should be visible.

pub fn show_vertical_scrollbar(self, flag: impl Res<bool> + 'static) -> Self

Sets whether the vertical scrollbar should be visible.

impl Handle<'_, Select>

pub fn placeholder<P: ToStringLocalized + Clone + 'static>( self, placeholder: impl Res<P> + 'static, ) -> Self

Sets the placeholder text that appears when the textbox has no value.

pub fn on_select<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when an option is selected.

pub fn min_selected(self, min_selected: impl Res<usize> + 'static) -> Self

Sets the minimum number of selected items.

pub fn max_selected(self, max_selected: impl Res<usize> + 'static) -> Self

Sets the maximum number of selected items.

impl Handle<'_, Spinbox>

pub fn on_change<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, f64),

Sets the callback triggered when the spinbox value is changed.

pub fn on_increment<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback which is triggered when the Spinbox value is incremented.

pub fn on_decrement<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback which is triggered when the Spinbox value is decremented.

pub fn vertical<U: Into<bool> + Clone + 'static>( self, vertical: impl Res<U> + 'static, ) -> Self

Sets the orientation of the Spinbox to vertical.

pub fn icons(self, icons: impl Res<SpinboxIcons> + 'static) -> Self

Set the icons which should be used for the increment and decrement buttons of the Spinbox

pub fn min<U: Into<f64> + Clone + 'static>( self, min: impl Res<U> + 'static, ) -> Self

Sets the minimum value of the Spinbox, disabling the decrement button when reached.

pub fn max<U: Into<f64> + Clone + 'static>( self, max: impl Res<U> + 'static, ) -> Self

Sets the maximum value of the Spinbox, disabling the increment button when reached.

impl Handle<'_, Switch>

pub fn on_toggle<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the Switch is pressed.

Examples

Switch::new(cx, value)
    .on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl Handle<'_, TabView>

pub fn vertical(self) -> Self

pub fn on_select( self, callback: impl Fn(&mut EventContext<'_>, usize) + 'static, ) -> Self

pub fn with_selected<U: Into<usize>>(self, selected: impl Res<U>) -> Self

impl<R, T> Handle<'_, Textbox<R, T>>
where R: Res<T> + 'static, T: Clone + ToStringLocalized + FromStr + 'static,

pub fn on_edit<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, String) + Send + Sync,

Sets the callback triggered when a textbox is edited, i.e. text is inserted/deleted.

Callback provides the current text of the textbox.

pub fn on_submit<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, T, bool) + Send + Sync,

Sets the callback triggered when a textbox is submitted, i.e. when the enter key is pressed with a single-line textbox or the textbox loses focus.

Callback provides the text of the textbox and a flag to indicate if the submit was due to a key press or a loss of focus.

pub fn on_blur<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox is blurred, i.e. the mouse is pressed outside of the textbox.

pub fn on_cancel<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox edit is cancelled, i.e. the escape key is pressed while editing.

pub fn validate<F>(self, is_valid: F) -> Self
where F: 'static + Fn(&T) -> bool + Send + Sync,

Sets a validation closure which is called when the textbox is edited and sets the validity attribute to the output of the closure.

If a textbox is modified with the validate modifier then the on_submit will not be called if the text is invalid.

pub fn placeholder<P: ToStringLocalized + Clone + 'static>( self, text: impl Res<P> + 'static, ) -> Self

Sets the placeholder text that appears when the textbox has no value.

impl Handle<'_, ToggleButton>

pub fn on_toggle( self, callback: impl Fn(&mut EventContext<'_>) + 'static, ) -> Self

Sets the callback triggered when the ToggleButton is toggled.

impl Handle<'_, Tooltip>

pub fn placement<U: Into<Placement> + Clone + 'static>( self, placement: impl Res<U> + 'static, ) -> Self

Sets the position where the tooltip should appear relative to its parent element. Defaults to Placement::Bottom.

pub fn arrow<U: Into<bool> + Clone + 'static>( self, show_arrow: impl Res<U> + 'static, ) -> Self

Sets whether the tooltip should include an arrow. Defaults to true.

pub fn arrow_size<U: Into<Length> + Clone + 'static>( self, size: impl Res<U> + 'static, ) -> Self

Sets the size of the tooltip arrow if enabled.

impl Handle<'_, VirtualList>

pub fn selection<R>(self, selection: impl Res<R> + 'static) -> Self
where R: Deref<Target = [usize]> + Clone + 'static,

Sets the selected items of the list from a signal of indices.

pub fn on_select<F>(self, callback: F) -> Self
where F: 'static + Fn(&mut EventContext<'_>, usize),

Sets the callback triggered when a ListItem is selected.

pub fn selectable<U: Into<Selectable> + Clone + 'static>( self, selectable: impl Res<U> + 'static, ) -> Self

Set the selectable state of the List.

pub fn selection_follows_focus<U: Into<bool> + Clone + 'static>( self, flag: impl Res<U> + 'static, ) -> Self

Sets whether the selection should follow the focus.

pub fn scroll_to_cursor(self, flag: bool) -> Self

Sets whether the scrollbar should move to the cursor when pressed.

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + 'static + Send + Sync, ) -> Self

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_x(self, scrollx: impl Res<f32> + 'static) -> Self

Set the horizontal scroll position of the ScrollView. Accepts a value or lens to an ‘f32’ between 0 and 1.

pub fn scroll_y(self, scrollx: impl Res<f32> + 'static) -> Self

Set the vertical scroll position of the ScrollView. Accepts a value or lens to an ‘f32’ between 0 and 1.

pub fn show_horizontal_scrollbar(self, flag: impl Res<bool> + 'static) -> Self

Sets whether the horizontal scrollbar should be visible.

pub fn show_vertical_scrollbar(self, flag: impl Res<bool> + 'static) -> Self

Sets whether the vertical scrollbar should be visible.

impl Handle<'_, XYPad>

pub fn on_change<F: Fn(&mut EventContext<'_>, f32, f32) + 'static>( self, callback: F, ) -> Self

Set the callback which will be triggered when the XYPad is manipulated.