vizia_core

Module views

Built-in views provided by vizia.

Structs

ArcTrack

Makes a knob that represents the current value with an arc

Avatar

An avatar view is used to visually represent a person or entity and can contain text, an icon, or an image.

AvatarGroup

The AvatarGroup view can be used to group a series of avatars together.

Badge

A Badge view for showing notifications, counts, or status information.

Button

A simple push button with a contained view.

ButtonGroup

A view which represents a group of buttons.

Checkbox

A checkbox used to display and toggle a boolean state.

Chip

A visual indicator such as a tag.

Collapsible

A collapsible view that can be opened or closed to hide content.

ComboBox

A ComboBox view which combines a textbox with a list popup, allowing users to filter options by typing.

Datepicker

A control used to select a date.

Divider

The Divider view provides a thin, unobtrusive line for visually separating views.

Dropdown

A dropdown is used to display some state with the ability to open a popup with options to change that state.

Element

A basic element with no interactivity.

Grid

A view which arranges its children into a grid.

HStack

A view which arranges its children into a horizontal stack (row).

Image

A view which presents an image.

Knob

A circular view which represents a value.

Label

A label used to display text.

List

A view for creating a list of items from an iterable signal.

ListItem

A view which represents a selectable item within a list.

Markdown

A view which parses and displays markdown as rich text.

MenuBar

A view which represents a horizontal group of menus.

MenuButton

A view which represents a pressable item within a menu.

PickList

A view which allows the user to select an item from a dropdown list.

Popup

A view for displaying popup content.

PopupData

A model which can be used by views which contain a popup.

ProgressBar

A simple progress bar that can be used to show the progress of something.

RadioButton

A radio button used to display and toggle boolean state.

Rating

A view which represents a rating as a number of filled stars.

ResizableStack

A view that can be resized by clicking and dragging from the right or bottom edge.

ResizeHandle

ScrollView

A container a view which allows the user to scroll any overflowed content.

Scrollbar

A view which represents a bar that can be dragged to manipulate a scrollview.

Slider

The slider control can be used to select from a continuous set of values.

SliderDataInternal

Internal data used by the slider.

Spacer

A view which can be used to add stretch space between two other views.

Spinbox

A view which represents a value which can be incremented or decremented.

Submenu

A view which represents a submenu within a menu.

Svg

A view which presents an SVG image.

Switch

A Switch used to display and toggle a boolean state.

TabHeader

TabPair

TabView

TextSpan

A view which represents a span of text within a label.

Textbox

The Textbox view provides an input control for editing a value as a string.

TickKnob

Makes a round knob with a tick to show the current value

Ticks

Adds tickmarks to a knob to show the steps that a knob can be set to. When added to a knob, the knob should be made smaller (depending on span), so the knob doesn’t overlap with the tick marks

ToggleButton

A button which can be toggled between two states.

Tooltip

A tooltip view.

VStack

A view which arranges its children into a vertical stack (column).

VirtualList

A view for creating a list of items from a binding to an iteratable list. Rather than creating a view for each item, items are recycled in the list.

XYPad

A view which allows the user to manipulate 2 floating point values simultaneously on a two dimensional pane.

ZStack

A view which overlays its children on top of each other.

Enums

AvatarVariant

Enum which represents the geometric variants of an avatar view.

BadgePlacement

Enum which represents the placement of a badge on its parent.

ButtonVariant

Used in conjunction with the variant modifier for selecting the style variant of a button or button group.

ChipVariant

Used in conjunction with the variant modifier for selecting the style variant of a chip.

CollapsibleEvent

Events that can be triggered by the collapsible view.

KnobMode

ListEvent

Events used by the List view

MenuEvent

Events used by menus.

Orientation

The orientation of a widget, such as a slider or scrollbar

Placement

Describes the placement of a view relative to its parent element.

PopupEvent

Events used by the Popup view.

ResizableStackEvent

Events emitted by the ResizableStack view to indicate changes in dragging state.

ResizeStackDirection

A direction for resizing a resizable stack, either horizontally (right) or vertically (bottom).

ScrollEvent

Events for setting the properties of a scroll view.

Selectable

Represents how items can be selected in a list.

SpinboxIcons

And enum which represents the icons that can be used for the increment and decrement buttons of the Spinbox.

TabEvent

TextEvent

Events for modifying a textbox.

Traits

ButtonModifiers

Modifiers for changing the appearance of buttons.