Function vizia_core::views::setup_menu_entry

pub fn setup_menu_entry<T, F1, F2>(
    handle: Handle<'_, T>,
    on_select: F1,
    on_deselect: F2
) -> Handle<'_, T>where
    T: View,
    F1: 'static + Fn(&mut Context),
    F2: 'static + Fn(&mut Context),

A helper function which sets up the necessary attributes on a view to be a menu entry. Call this with a handle to an object you would like to be considered a menu entry. It adds an on_over event handler updating the current selected menu entry and binds to said selection, updating the selected pseudo-class accordingly and calling the on_select and on_deselect callbacks appropriately.