Crate vizia_core

Modules

animation

Animations

cache

The cache is a store for intermediate data produced while computing state, notably layout results. The main type here is CachedData, usually accessed via cx.cache.

context

entity

environment

events

Events

fonts

handle

image

This is a re-export of image.

input

layout

Layout

localization

modifiers

Methods on views for changing their style properties or for adding actions.

prelude

Members which we recommend you wildcard-import.

resource

state

Data Binding

style

systems

text

tree

vg

This is a re-export of femtovg.

view

views

This is the module where all types which implement View live.

window

Functions

log

One very small function for abstracting debugging between web and desktop programming. On the desktop, it will print to stdout, and on the web, it will print to the console log.