Module vizia_core::context

Modules

backend

Structs

Context

The main storage and control object for a Vizia application.

ContextProxy

A bundle of data representing a snapshot of the context when a thread was spawned.

DrawCache

Cached data used for drawing.

DrawContext

A restricted context used when drawing.

EventContext

Enums

ProxyEmitError

Errors that might come up when emitting an event via a ContextProxy.

Traits

DataContext

A trait for any Context-like object that lets you access stored model data.

EventProxy