Module vizia_core::context

Context types for retained state, used during view building, event handling, and drawing.

Structs

AccessContext

A context used for configuring the accessibility features of a view.

AccessNode

Wrapper around an accesskit node builder, a node id, and a list of children to be added to the node.

Context

The main storage and control object for a Vizia application.

ContextProxy

A bundle of data representing a snapshot of the context when a thread was spawned.

DrawContext

A context used when drawing.

EventContext

A context used when handling events.

ResourceContext

A context used when loading images.

Enums

ProxyEmitError

Errors that might occur when emitting an event via a ContextProxy.

Constants

CURRENT

MAPS

MAP_MANAGER

Traits

DataContext

A trait for any Context-like object that lets you access stored model data.

EmitContext

EventProxy

TreeProps

Trait for querying properties of the tree from a context.