Struct vizia_core::context::DrawContext

pub struct DrawContext<'a> { /* private fields */ }

A context used when drawing.

The DrawContext is provided by the draw method in View and can be used to immutably access the computed style and layout properties of the current view.

Example

pub struct CustomView {}

impl CustomView {
    pub fn new(cx: &mut Context) -> Handle<Self> {
        Self{}.build(cx, |_|{})
    }
}

impl View for CustomView {
    fn draw(&self, cx: &mut DrawContext, canvas: &mut Canvas) {
        // Get the computed bounds after layout of the current view
        let bounds = cx.bounds();
        // Draw to the canvas using the bounds of the current view
        let mut path = vg::Path::new();
        path.rect(bounds.x, bounds.y, bounds.w, bounds.h);
        canvas.fill_path(&mut path, &vg::Paint::color(vg::Color::rgb(200, 100, 100)));
    }
}

Implementations

impl<'a> DrawContext<'a>

pub fn bounds(&self) -> BoundingBox

Returns the bounds of the current view.

pub fn scale_factor(&self) -> f32

Returns the scale factor.

pub fn modifiers(&self) -> &Modifiers

Returns a reference to the keyboard modifiers state.

pub fn mouse(&self) -> &MouseState<Entity>

Returns a reference to the mouse state.

pub fn clip_region(&self) -> BoundingBox

Returns the bounding box of the clip region of the current view.

pub fn transform(&self) -> Transform2D

Returns the 2D transform of the current view.

pub fn visibility(&self) -> Option<Visibility>

Returns the visibility of the current view.

pub fn display(&self) -> Display

Returns the display of the current view.

pub fn opacity(&self) -> f32

Returns the opacity of the current view.

pub fn default_font(&self) -> &[FamilyOwned]

Returns the lookup pattern to pick the default font.

pub fn font_size(&self, entity: Entity) -> f32

Returns the font-size of the current view in physical pixels.

pub fn logical_to_physical(&self, logical: f32) -> f32

Function to convert logical points to physical pixels.

pub fn physical_to_logical(&self, physical: f32) -> f32

Function to convert physical pixels to logical points.

pub fn border_width(&self) -> f32

Returns the border width of the current view in physical pixels.

pub fn outline_color(&self) -> Color

Returns the outline color of the current view.

pub fn outline_width(&self) -> f32

Returns the outline width of the current view in physical pixels.

pub fn outline_offset(&self) -> f32

Returns the outline offset of the current view in physcial pixels.

pub fn border_top_left_radius(&self) -> f32

Returns the border radius for the top-left corner of the current view.

pub fn border_top_right_radius(&self) -> f32

Returns the border radius for the top-right corner of the current view.

pub fn border_bottom_left_radius(&self) -> f32

Returns the border radius for the bottom-left corner of the current view.

pub fn border_bottom_right_radius(&self) -> f32

Returns the border radius for the bottom-right corner of the current view.

pub fn border_top_left_shape(&self) -> BorderCornerShape

Returns the border corner shape for the top-left corner of the current view.

pub fn border_top_right_shape(&self) -> BorderCornerShape

Returns the border corner shape for the top-left corner of the current view.

pub fn border_bottom_left_shape(&self) -> BorderCornerShape

Returns the border corner shape for the top-left corner of the current view.

pub fn border_bottom_right_shape(&self) -> BorderCornerShape

Returns the border corner shape for the top-left corner of the current view.

pub fn child_left(&self) -> Units

Returns the child-left space of the current view.

pub fn child_right(&self) -> Units

Returns the child-right space of the current view.

pub fn child_top(&self) -> Units

Returns the child-top space of the current view.

pub fn child_bottom(&self) -> Units

Returns the child-bottom space of the current view.

pub fn background_color(&self) -> Color

pub fn border_color(&self) -> Color

pub fn selection_color(&self) -> Color

pub fn caret_color(&self) -> Color

pub fn font_color(&self) -> Color

pub fn text_wrap(&self) -> bool

Returns whether the current view should have its text wrapped.

pub fn text_align(&self) -> Option<TextAlign>

pub fn box_shadows(&self) -> Option<&Vec<BoxShadow>>

pub fn backdrop_filter(&self) -> Option<&Filter>

pub fn background_images(&self) -> Option<&Vec<ImageOrGradient>>

pub fn background_size(&self) -> Vec<BackgroundSize>

pub fn build_path(&mut self) -> Path

Get the vector path of the current view.

pub fn draw_backdrop_filter(&mut self, canvas: &mut Canvas, path: &mut Path)

Draw backdrop filters for the current view.

pub fn draw_background(&mut self, canvas: &mut Canvas, path: &mut Path)

Draw background color or background image (including gradients) for the current view.

pub fn draw_text_and_selection(&mut self, canvas: &mut Canvas)

pub fn draw_border(&mut self, canvas: &mut Canvas, path: &mut Path)

Draw the border of the current view.

pub fn draw_outline(&mut self, canvas: &mut Canvas)

Draw the outline of the current view.

pub fn draw_inset_box_shadows(&mut self, canvas: &mut Canvas, path: &mut Path)

Draw inset box shadows for the current view.

pub fn draw_shadows(&mut self, canvas: &mut Canvas, path: &mut Path)

Draw non-inset box shadows for the current view.

pub fn draw_text( &mut self, canvas: &mut Canvas, bounds: BoundingBox, justify: (f32, f32) )

Draw any text for the current view.

pub fn draw_text_selection( &mut self, canvas: &mut Canvas, bounds: BoundingBox, justify: (f32, f32) )

Draw the selection box for the text of the current view.

pub fn draw_text_caret( &mut self, canvas: &mut Canvas, bounds: BoundingBox, justify: (f32, f32), width: f32 )

Draw text caret for the current view.