Module vizia_core::events

Events

Views communicate with each other and model data via events. An Event contains a Message, as well as metadata to describe how events should propagate through the tree. By default events will propagate up the tree from the target.

A Message can be any static type but is usually an enum. For example:

enum MyEvent {
    ReadDocs,
    CloseDocs,    
}

Then, to send an event up the tree from the current entity:

cx.emit(MyEvent::ReadDocs);

Or, to send an event from the current entity directly to a target:

cx.emit_to(target, MyEvent::ReadDocs);

Views and Models receive events through the event() method of the View or Model traits. The event message must then be downcast to the right type:

fn on_event(&mut self, state: &mut State, entity: Entity, event: &mut Event) {
    event.map(|my_event, _| match my_event {
        MyEvent::ReadDocs => {
            // Do something
        }

        MyEvent::CloseDocs => {
            // Do something else
        }
    });
}

Structs

Event

A wrapper around a message, providing metadata on how the event travels through the tree.

EventMeta

The meta data of an Event.

Enums

Propagation

Determines how an event propagates through the tree.

Traits

Message

The content of an event.

ViewHandler