vizia_core::localization

Trait ToStringLocalized

pub trait ToStringLocalized {
    // Required method
    fn to_string_local(&self, cx: &impl DataContext) -> String;
}

A trait for converting from Localized to a String via a translation using fluent.

Required Methods

fn to_string_local(&self, cx: &impl DataContext) -> String

Method for converting the current type to a String via a translation using fluent.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.