Module vizia_core::modifiers

Methods on views for changing their style properties or for adding actions.

Structs

EventHandle

Traits

AbilityModifiers

Modifiers for changing the abilities of a view.

ActionModifiers

Modifiers which add an action callback to a view.

LayoutModifiers

Modifiers for changing the layout properties of a view.

StyleModifiers

Modifiers for changing the style properties of a view.

TextModifiers

Modifiers for changing the text properties of a view.