Trait vizia_core::modifiers::StyleModifiers

pub trait StyleModifiers: Modifiable {

    fn id(self, id: impl Into<String>) -> Self { ... }
    fn class(self, name: &str) -> Self { ... }
    fn toggle_class(self, name: &str, applied: impl Res<bool>) -> Self { ... }
    fn checked<U: Into<bool>>(self, state: impl Res<U>) -> Self { ... }
    fn disabled<U: Into<bool>>(self, value: impl Res<U>) -> Self { ... }
    fn display<U: Into<Display>>(self, value: impl Res<U>) -> Self { ... }
    fn visibility<U: Into<Visibility>>(self, value: impl Res<U>) -> Self { ... }
    fn z_order<U: Into<i32>>(self, value: impl Res<U>) -> Self { ... }
    fn overflow<U: Into<Overflow>>(self, value: impl Res<U>) -> Self { ... }
    fn background_color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn background_image<U: Into<String>>(self, value: impl Res<U>) -> Self { ... }
    fn image<U: ToString>(self, value: impl Res<U>) -> Self { ... }
    fn border_width<U: Into<Units>>(self, value: impl Res<U>) -> Self { ... }
    fn border_color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn border_radius_top_left<U: Into<Units>>(self, value: impl Res<U>) -> Self { ... }
    fn border_radius_top_right<U: Into<Units>>(self, value: impl Res<U>) -> Self { ... }
    fn border_radius_bottom_left<U: Into<Units>>(
        self,
        value: impl Res<U>
    ) -> Self { ... }
    fn border_radius_bottom_right<U: Into<Units>>(
        self,
        value: impl Res<U>
    ) -> Self { ... }
    fn border_radius<U: Into<Units>>(self, value: impl Res<U>) -> Self { ... }
    fn border_shape_top_left<U: Into<BorderCornerShape>>(
        self,
        value: impl Res<U>
    ) -> Self { ... }
    fn border_shape_top_right<U: Into<BorderCornerShape>>(
        self,
        value: impl Res<U>
    ) -> Self { ... }
    fn border_shape_bottom_left<U: Into<BorderCornerShape>>(
        self,
        value: impl Res<U>
    ) -> Self { ... }
    fn border_shape_bottom_right<U: Into<BorderCornerShape>>(
        self,
        value: impl Res<U>
    ) -> Self { ... }
    fn border_corner_shape<U: Into<BorderCornerShape>>(
        self,
        value: impl Res<U>
    ) -> Self { ... }
    fn outline_width<U: Into<Units>>(self, value: impl Res<U>) -> Self { ... }
    fn outline_color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn outline_offset<U: Into<Units>>(self, value: impl Res<U>) -> Self { ... }
    fn cursor<U: Into<CursorIcon>>(self, value: impl Res<U>) -> Self { ... }
    fn rotate<U: Into<f32>>(self, value: impl Res<U>) -> Self { ... }
    fn translate<U: Into<(f32, f32)>>(self, value: impl Res<U>) -> Self { ... }
    fn scale<U: Into<(f32, f32)>>(self, value: impl Res<U>) -> Self { ... }

}

Modifiers for changing the style properties of a view.

Provided Methods

fn id(self, id: impl Into<String>) -> Self

Sets the ID name of the view.

The ID name can be references by a CSS selector.

Example

Element::new(cx).id("foo");

css

#foo {
    background-color: red;
}

fn class(self, name: &str) -> Self

Adds a class name to the view.

fn toggle_class(self, name: &str, applied: impl Res<bool>) -> Self

Sets whether a view should have the given class name.

fn checked<U: Into<bool>>(self, state: impl Res<U>) -> Self

Sets the state of the view to checked.

fn disabled<U: Into<bool>>(self, value: impl Res<U>) -> Self

Sets the view to be disabled.

This property is inherited by the descendants of the view.

fn display<U: Into<Display>>(self, value: impl Res<U>) -> Self

Sets whether the view should be positioned and rendered.

A display value of Display::None causes the view to be ignored by both layout and rendering.

fn visibility<U: Into<Visibility>>(self, value: impl Res<U>) -> Self

Sets whether the view should be rendered.

The layout system will still compute the size and position of an invisible view.

fn z_order<U: Into<i32>>(self, value: impl Res<U>) -> Self

Sets the z-order index of the view.

Views with a higher z-order will be rendered on top of those with a lower z-order. Views with the same z-order are rendered in tree order.

fn overflow<U: Into<Overflow>>(self, value: impl Res<U>) -> Self

Sets the overflow behavior of the view.

The overflow behavior determines whether child views can render outside the bounds of their parent.

fn background_color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the background color of the view.

fn background_image<U: Into<String>>(self, value: impl Res<U>) -> Self

Sets the background image of the view.

Background image will override any background gradient or color.

fn image<U: ToString>(self, value: impl Res<U>) -> Self

fn border_width<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the border width of the view.

fn border_color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the border color of the view.

fn border_radius_top_left<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the border radius for the top-left corner of the view.

fn border_radius_top_right<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the border radius for the top-right corner of the view.

fn border_radius_bottom_left<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the border radius for the bottom-left corner of the view.

fn border_radius_bottom_right<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the border radius for the bottom-right corner of the view.

fn border_radius<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the border radius for all four corners of the view.

fn border_shape_top_left<U: Into<BorderCornerShape>>(
self,
value: impl Res<U>
) -> Self

Sets the border corner shape for the top-left corner of the view.

fn border_shape_top_right<U: Into<BorderCornerShape>>(
self,
value: impl Res<U>
) -> Self

Sets the border corner shape for the top-right corner of the view.

fn border_shape_bottom_left<U: Into<BorderCornerShape>>(
self,
value: impl Res<U>
) -> Self

Sets the border corner shape for the bottom-left corner of the view.

fn border_shape_bottom_right<U: Into<BorderCornerShape>>(
self,
value: impl Res<U>
) -> Self

Sets the border corner shape for the bottom-right corner of the view.

fn border_corner_shape<U: Into<BorderCornerShape>>(
self,
value: impl Res<U>
) -> Self

Sets the border corner shape for all four corners of the view.

fn outline_width<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the outline width of the view.

fn outline_color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the outline color of the view.

fn outline_offset<U: Into<Units>>(self, value: impl Res<U>) -> Self

Sets the outline offset of the view.

fn cursor<U: Into<CursorIcon>>(self, value: impl Res<U>) -> Self

Sets the mouse cursor used when the view is hovered.

fn rotate<U: Into<f32>>(self, value: impl Res<U>) -> Self

Sets the angle of rotation for the view.

Rotation applies to the rendered view and does not affect layout.

fn translate<U: Into<(f32, f32)>>(self, value: impl Res<U>) -> Self

Sets the translation offset of the view.

Translation applies to the rendered view and does not affect layout.

fn scale<U: Into<(f32, f32)>>(self, value: impl Res<U>) -> Self

Sets the scale of the view.

Scale applies to the rendered view and does not affect layout.