Trait vizia_core::modifiers::TextModifiers

pub trait TextModifiers: Modifiable {
    fn text<U: ToString>(self, value: impl Res<U>) -> Self { ... }
    fn font_family<U: Into<Vec<FamilyOwned>>>(self, value: impl Res<U>) -> Self { ... }
    fn font_weight<U: Into<Weight>>(self, value: impl Res<U>) -> Self { ... }
    fn font_style<U: Into<FontStyle>>(self, value: impl Res<U>) -> Self { ... }
    fn color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn font_size(self, value: impl Res<f32>) -> Self { ... }
    fn caret_color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn selection_color<U: Into<Color>>(self, value: impl Res<U>) -> Self { ... }
    fn text_wrap<U: Into<bool>>(self, value: impl Res<U>) -> Self { ... }
}

Modifiers for changing the text properties of a view.

Provided Methods

fn text<U: ToString>(self, value: impl Res<U>) -> Self

Sets the text content of the view.

fn font_family<U: Into<Vec<FamilyOwned>>>(self, value: impl Res<U>) -> Self

Sets the font that should be used by the view.

The font name refers to the name assigned when the font is added to context.

fn font_weight<U: Into<Weight>>(self, value: impl Res<U>) -> Self

Sets the font weight that should be used by the view.

fn font_style<U: Into<FontStyle>>(self, value: impl Res<U>) -> Self

Sets the font style that should be used by the view.

fn color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the text color of the view.

fn font_size(self, value: impl Res<f32>) -> Self

Sets the font size of the view.

fn caret_color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the ext caret color of the view.

fn selection_color<U: Into<Color>>(self, value: impl Res<U>) -> Self

Sets the color used to highlight selected text within the view.

fn text_wrap<U: Into<bool>>(self, value: impl Res<U>) -> Self

Sets whether the text of the view should be allowed to wrap.