Struct vizia_core::prelude::Tree

pub struct Tree<I>where
    I: GenerationalId,{
    pub parent: Vec<Option<I>, Global>,
    pub first_child: Vec<Option<I>, Global>,
    pub next_sibling: Vec<Option<I>, Global>,
    pub prev_sibling: Vec<Option<I>, Global>,
    pub ignored: Vec<bool, Global>,
    pub lock_focus_within: Vec<bool, Global>,
    pub changed: bool,
}

The Tree describes the tree of entities.

This type is part of the prelude.

Fields
parent: Vec<Option<I>, Global>first_child: Vec<Option<I>, Global>next_sibling: Vec<Option<I>, Global>prev_sibling: Vec<Option<I>, Global>ignored: Vec<bool, Global>lock_focus_within: Vec<bool, Global>changed: bool
Implementations

impl<I> Tree<I>where
I: GenerationalId,

pub fn new() -> Tree<I>

Creates a new tree with a root entity.

pub fn get_child_index(&self, entity: I) -> Option<usize>

pub fn get_last_child(&self, entity: I) -> Option<I>

Returns the last child of an entity.

pub fn get_child(&self, entity: I, n: usize) -> Option<I>

Returns the nth child of an entity.

pub fn get_num_children(&self, entity: I) -> Option<u32>

Returns the number of children of an entity.

pub fn is_ignored(&self, entity: I) -> bool

Returns true if the node should be skipped by layout

pub fn lock_focus_within(&self, entity: I) -> I

Returns the first ancestor to have the lock_focus_within flag set

pub fn get_layout_parent(&self, entity: I) -> Option<I>

Returns the first ancestor of an entity which is not ignored

pub fn get_parent(&self, entity: I) -> Option<I>

Returns the parent of an entity.

pub fn get_first_child(&self, entity: I) -> Option<I>

Returns the first child of an entity or None if there isn’t one.

pub fn get_next_sibling(&self, entity: I) -> Option<I>

Returns the next sibling of an entity or None if t here isn’t one.

pub fn get_prev_sibling(&self, entity: I) -> Option<I>

Returns the previous sibling of an entity or None if there isn’t one.

pub fn is_first_child(&self, entity: I) -> bool

Returns true if the entity is the first child of its parent.

pub fn is_last_child(&self, entity: I) -> bool

Returns true if the entity is the last child of its parent.

pub fn is_sibling(&self, entity1: I, entity2: I) -> bool

pub fn has_children(&self, entity: I) -> bool

Returns true if the entity has children.

pub fn remove(&mut self, entity: I) -> Result<(), TreeError>

Removes an entity from the tree

This method assumes that a check if the entity is alive has already been done prior to calling this method.

pub fn set_first_child(&mut self, entity: I) -> Result<(), TreeError>

Makes the entity the first child of its parent.

pub fn set_next_sibling(&mut self, entity: I, sibling: I) -> Result<(), TreeError>

pub fn set_prev_sibling(&mut self, entity: I, sibling: I) -> Result<(), TreeError>

pub fn set_parent(&mut self, entity: I, parent: I)

pub fn set_ignored(&mut self, entity: I, flag: bool)

pub fn set_lock_focus_within(&mut self, entity: I, flag: bool)

pub fn add(&mut self, entity: I, parent: I) -> Result<(), TreeError>

Adds an entity to the tree with the specified parent.