Trait vizia_core::vg::rgb::ComponentSlice

pub trait ComponentSlice<T> {
    // Required methods
    fn as_slice(&self) -> &[T];
    fn as_mut_slice(&mut self) -> &mut [T];
}

Casting the struct to slices of its components

Required Methods

fn as_slice(&self) -> &[T]

The components interpreted as an array, e.g. one RGB expands to 3 elements.

It’s implemented for individual pixels as well as slices of pixels.

fn as_mut_slice(&mut self) -> &mut [T]

The components interpreted as a mutable array, e.g. one RGB expands to 3 elements.

It’s implemented for individual pixels as well as slices of pixels.

If you get an error when calling this on an array, add [..]

use of unstable library feature ‘array_methods’

arr[..].as_mut_slice()