Struct vizia_core::view::Handle

pub struct Handle<'a, V> { /* private fields */ }

A handle to a view which has been already built into the tree.

Implementations

impl<'a, V> Handle<'a, V>

pub fn entity(&self) -> Entity

Returns the [Entity] id of the view.

pub fn context(&mut self) -> &mut Context

Returns a mutable reference to the context.

pub fn parent(&self) -> Entity

pub fn lock_focus_to_within(self) -> Self

Stop the user from tabbing out of a subtree, which is useful for modal dialogs.

pub fn modify<F>(self, f: F) -> Selfwhere F: FnOnce(&mut V), V: 'static,

Mody the internal data of the view.

pub fn on_build<F>(self, callback: F) -> Selfwhere F: Fn(&mut EventContext<'_>),

Callback which is run when the view is built/rebuilt.

pub fn bind<L, F>(self, lens: L, closure: F) -> Selfwhere L: Lens, <L as Lens>::Target: Data, F: 'static + Fn(Handle<'_, V>, L),

pub fn needs_relayout(&mut self)

Marks the view as needing a relayout.

pub fn needs_restyle(&mut self)

Marks the view as needing a restyle.

pub fn needs_redraw(&mut self)

Marks the view as needing a redraw.

pub fn bounds(&self) -> BoundingBox

Returns the bounding box of the view.

pub fn scale_factor(&self) -> f32

Returns the scale factor of the device.

impl Handle<'_, Checkbox>

pub fn on_toggle<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the checkbox is pressed.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, Chip>

pub fn on_close( self, callback: impl 'static + Fn(&mut EventContext<'_>) + Send + Sync ) -> Self

impl<'v, L1, L2, T> Handle<'v, ComboBox<L1, L2, T>>where L1: Lens<Target = Vec<T>>, T: 'static + Data + ToString, L2: Lens<Target = usize>,

pub fn on_select<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, usize),

impl<'a> Handle<'a, Datepicker>

pub fn on_select<F: 'static + Fn(&mut EventContext<'_>, NaiveDate)>( self, callback: F ) -> Self

impl<'a, L: Lens<Target = f32>> Handle<'a, Knob<L>>

pub fn on_changing<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, f32),

impl Handle<'_, ArcTrack>

pub fn value<L: Lens<Target = f32>>(self, lens: L) -> Self

impl Handle<'_, TickKnob>

pub fn value<L: Lens<Target = f32>>(self, lens: L) -> Self

impl Handle<'_, Label>

pub fn describing(self, entity_identifier: impl Into<String>) -> Self

Which form element does this label describe.

Examples

Checkbox::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue)).id("checkbox_identifier");
Label::new(cx, "hello").describing("checkbox_identifier");

impl<L: Lens<Target = Vec<T>>, T: Data> Handle<'_, List<L, T>>

pub fn on_increment<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

pub fn on_decrement<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

pub fn on_clear<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

impl<'a, L> Handle<'a, MenuPopup<L>>where L: Lens, L::Target: Clone + Into<bool>,

pub fn on_blur<F>(self, f: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl<'v> Handle<'v, PickList>

pub fn on_select<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, usize),

impl<'a, L> Handle<'a, Popup<L>>where L: Lens, L::Target: Clone + Into<bool>,

pub fn on_blur<F>(self, f: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

Registers a callback for when the user clicks off of the popup, usually with the intent of closing it.

impl<'a> Handle<'a, ProgressBar>

pub fn bar_color(self, color: impl Res<Color>) -> Self

Set the color of the bar inside the ProgressBar.

you also pass a lens to this method if you want to be able to change the color dynamically.

impl Handle<'_, RadioButton>

pub fn on_select<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the radio button is selected.

Examples

RadioButton::new(cx, AppData::value).on_select(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, Rating>

pub fn on_change<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, u32),

impl<'a, L1: 'static + Lens<Target = f32>> Handle<'a, Scrollbar<L1>>

pub fn scroll_to_cursor(self, scroll_to_cursor: impl Res<bool>) -> Self

impl<'a, L: Lens> Handle<'a, ScrollView<L>>

pub fn on_scroll( self, callback: impl Fn(&mut EventContext<'_>, f32, f32) + 'static + Send + Sync ) -> Self

Sets a callback which will be called when a scrollview is scrolled, either with the mouse wheel, touchpad, or using the scroll bars.

pub fn scroll_to_cursor(self, scroll_to_cursor: bool) -> Self

impl<L: Lens> Handle<'_, Slider<L>>

pub fn on_changing<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, f32),

Sets the callback triggered when the slider value is changing (dragging).

Takes a closure which triggers when the slider value is changing, either by pressing the track or dragging the thumb along the track.

Example

Slider::new(cx, AppData::value)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

pub fn range(self, range: Range<f32>) -> Self

Sets the range of the slider.

If the bound data is outside of the range then the slider will clip to min/max of the range.

Example

Slider::new(cx, AppData::value)
    .range(-20.0..50.0)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

pub fn step(self, step: f32) -> Self

pub fn keyboard_fraction(self, keyboard_fraction: f32) -> Self

Sets the fraction of a slider that a press of an arrow key will change.

Example

Slider::new(cx, AppData::value)
    .keyboard_fraction(0.05)
    .on_changing(|cx, value| {
        println!("Slider on_changing: {}", value);
    });

impl Handle<'_, NamedSlider>

pub fn on_changing<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, f32),

pub fn range(self, range: Range<f32>) -> Self

pub fn keyboard_fraction(self, keyboard_fraction: f32) -> Self

impl<'a> Handle<'a, Spinbox>

pub fn on_increment<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

pub fn on_decrement<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

impl Handle<'_, Switch>

pub fn on_toggle<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>),

Set the callback triggered when the Switch is pressed.

Examples

Switch::new(cx, AppData::value).on_toggle(|cx| cx.emit(AppEvent::ToggleValue));

impl<'a> Handle<'a, TabView>

pub fn on_select( self, callback: impl Fn(&mut EventContext<'_>, usize) + 'static ) -> Self

pub fn with_selected<U: Into<usize>>(self, selected: impl Res<U>) -> Self

impl<'a, L: Lens> Handle<'a, Textbox<L>>

pub fn on_edit<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, String) + Send + Sync,

Sets the callback triggered when a textbox is edited, i.e. text is inserted/deleted.

Callback provides the current text of the textbox.

pub fn on_submit<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, L::Target, bool) + Send + Sync,

Sets the callback triggered when a textbox is submitted, i.e. when the enter key is pressed with a single-line textbox or the textbox loses focus.

Callback provides the text of the textbox and a flag to indicate if the submit was due to a key press or a loss of focus.

pub fn on_blur<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox is blurred, i.e. the mouse is pressed outside of the textbox.

pub fn on_cancel<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>) + Send + Sync,

Sets the callback triggered when a textbox edit is cancelled, i.e. the escape key is pressed while editing.

pub fn validate<F>(self, is_valid: F) -> Selfwhere F: 'static + Fn(&L::Target) -> bool + Send + Sync,

Sets a validation closure which is called when the textbox is edited and sets the validity attribute to the output of the closure.

If a textbox is modified with the validate modifier then the on_submit will not be called if the text is invalid.

pub fn placeholder<P: ToString>(self, text: impl Res<P>) -> Self

Sets the placeholder text that appears when the textbox has no value.

impl<'a> Handle<'a, Timepicker>

pub fn on_change<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, NaiveTime),

impl<'a, L, T> Handle<'a, DigitalTimepicker<L, T>>where L: Lens<Target = T>, T: Timelike + Data,

pub fn on_change<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, NaiveTime),

impl<'v, L, T> Handle<'v, AnalogTimepicker<L, T>>where L: Lens<Target = T>, T: Timelike + Data + Copy,

pub fn on_change<F>(self, callback: F) -> Selfwhere F: 'static + Fn(&mut EventContext<'_>, NaiveTime),

pub fn show_controls(self, value: bool) -> Self

pub fn change_page_on_select(self, value: bool) -> Self

impl<'a> Handle<'a, VirtualList>

pub fn scroll_to_cursor(self, flag: bool) -> Self

pub fn on_change( self, callback: impl Fn(&mut EventContext<'_>, Range<usize>) + 'static ) -> Self