Trait vizia_core::view::View

pub trait View: 'static + Sized {
    fn body(&mut self, cx: &mut Context) { ... }

    fn build<F>(self, cx: &mut Context, content: F) -> Handle<'_, Self>
    where
        F: FnOnce(&mut Context),
    { ... }

    fn element(&self) -> Option<&'static str> { ... }

    fn event(&mut self, cx: &mut Context, event: &mut Event) { ... }

    fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas) { ... }
}

A view is any object which can be displayed on the screen.

This trait is part of the prelude.

Provided methods

fn body(&mut self, cx: &mut Context)

fn build<F>(self, cx: &mut Context, content: F) -> Handle<'_, Self> where
F: FnOnce(&mut Context),

fn element(&self) -> Option<&'static str>

The name of the view. This is used in css: to style every single one of a given view, you specify the element name.

fn event(&mut self, cx: &mut Context, event: &mut Event)

fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas)