Module vizia_core::views

This is the module where all types which implement View live.

Every member of this module is part of the prelude.

Modules

normalized_map

Structs

ArcTrack

Makes a knob that represents the current value with an arc

Button

A simple push button with an action and views inside of it.

Checkbox

A checkbox used to display and toggle boolean state.

Dropdown

A dropdown is used to display some state with the ability to open a popup with options to change that state.

Element

A basic element with no interactivity.

HStack

A view which arranges its children into a horizontal stack (row).

Image

Knob

Label

A label used to display text to the screen.

List

A view for creating a list of items from a binding to a Vec

Menu

A button containing a menu when you click/hover it.

MenuButton

A MenuButton is an entry in a menu that can be clicked to perform some action. It has various constructors depending on whether you want to make this button show a check icon conditionally.

MenuController

A MenuController is a container object which holds a menu. It is responsible for managing the focus of the menu, i.e. grabbing click events until the menu is closed.

MenuStack

A MenuStack is a stack of views which can be menu entries. The only interesting thing about it is that it builds a MenuData into itself.

Popup

PopupData

RadioButton

ScrollData

ScrollView

Scrollbar

Slider

The slider control can be used to select from a continuous set of values.

Table

TableColumn

Textbox

TickKnob

Makes a round knob with a tick to show the current value

Ticks

Adds tickmarks to a knob to show the steps that a knob can be set to. When added to a knob, the knob should be made smaller (depending on span), so the knob doesn’t overlap with the tick marks

VStack

A view which arranges its children into a vertical stack (column).

ZStack

A view which overlays its children on top of each other.

Enums

KnobMode

MenuEvent

Menu control events.

Orientation

The orientation of a widget, such as a slider or scrollbar

PopupEvent

ScrollEvent

TextEvent

Functions

setup_menu_entry

A helper function which sets up the necessary attributes on a view to be a menu entry. Call this with a handle to an object you would like to be considered a menu entry. It adds an on_over event handler updating the current selected menu entry and binds to said selection, updating the selected pseudo-class accordingly and calling the on_select and on_deselect callbacks appropriately.